+++ modules/shared/core/src/sap/commerce/toolset/spring/SpringHelper.kt
@@ -25,13 +25,11 @@ object SpringHelper {
 
     fun resolveBeanDeclaration(element: PsiElement, beanId: String) = SpringService.EP.extensionList
         .firstNotNullOfOrNull { it.resolveBeanDeclaration(element, beanId) }
-        ?: SimpleSpringService.getService(element.project)
-            ?.resolveBeanDeclaration(element, beanId)
+        ?: SimpleSpringService.getService(element.project).resolveBeanDeclaration(element, beanId)
 
 
     fun resolveBeanClass(element: PsiElement, beanId: String): PsiClass? = SpringService.EP.extensionList
         .firstNotNullOfOrNull { it.resolveBeanClass(element, beanId) }
-        ?: SimpleSpringService.getService(element.project)
-            ?.resolveBeanClass(element, beanId)
+        ?: SimpleSpringService.getService(element.project).resolveBeanClass(element, beanId)
 
 }