+++ modules/spring/core/src/sap/commerce/toolset/spring/AdvancedSpringService.kt
@@ -24,19 +24,24 @@ import com.intellij.openapi.module.ModuleUtilCore
 import com.intellij.psi.PsiElement
 import com.intellij.spring.SpringManager
 import com.intellij.spring.model.utils.SpringModelSearchers
+import sap.commerce.toolset.Plugin
 import sap.commerce.toolset.project.descriptor.ModuleDescriptorType
 import sap.commerce.toolset.project.facet.YFacet
 
 class AdvancedSpringService : SpringService {
 
-    override fun resolveBeanDeclaration(element: PsiElement, beanId: String) = guessModule(element)
-        ?.let { springResolveBean(it, beanId) }
-        ?.springBean
-        ?.xmlTag
+    override fun resolveBeanDeclaration(element: PsiElement, beanId: String) = Plugin.SPRING.ifActive {
+        guessModule(element)
+            ?.let { springResolveBean(it, beanId) }
+            ?.springBean
+            ?.xmlTag
+    }
 
-    override fun resolveBeanClass(element: PsiElement, beanId: String) = guessModule(element)
-        ?.let { springResolveBean(it, beanId) }
-        ?.beanClass
+    override fun resolveBeanClass(element: PsiElement, beanId: String) = Plugin.SPRING.ifActive {
+        guessModule(element)
+            ?.let { springResolveBean(it, beanId) }
+            ?.beanClass
+    }
 
     private fun guessModule(element: PsiElement) = ModuleUtilCore.findModuleForPsiElement(element)
         ?: ModuleManager.getInstance(element.project)